// fem-vision-studio/design-mockup/src/lib/types/concrete/MnKappaDiagram.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FailureMode } from "./FailureMode";
import type { MnKappaPoint } from "./MnKappaPoint";

/**
 * Het M-κ-diagram bij vaste N, tot bezwijken.
 */
export type MnKappaDiagram = { 
/**
 * Normaalkracht waarbij het diagram is bepaald (kN, trek positief).
 */
n_kn: number, 
/**
 * Richting van het moment waarvoor is gerekend (+1: trek onder; −1: trek boven).
 */
moment_sign: number, 
/**
 * Punten van κ = 0 tot κ_u, in de richting `moment_sign` (M ≥ 0).
 */
points: Array<MnKappaPoint>, 
/**
 * Kromming en moment bij bezwijken.
 */
kappa_u_per_m: number, m_u_knm: number, 
/**
 * Grootste moment op het diagram (de momentweerstand M_Rd bij deze N).
 */
m_max_knm: number, 
/**
 * Kromming en moment waarbij de eerste wapeningslaag gaat vloeien (ε ≥ ε_yd).
 */
kappa_y_per_m: number | null, m_y_knm: number | null, failure_mode: FailureMode, 
/**
 * Drukzonehoogte en rekken bij bezwijken.
 */
x_u_mm: number, eps_c_u: number, eps_s_u: number, n_strips: number, };